feat: Binary syscall bypass through VDSO
Glibc utilizes the `__syscall` macro to perform system calls.
Syscalls with varying numbers of parameters are further dispatched to
implementations like `__syscall1`, `__syscall2`, and so on.
These implementations employ the `syscall` instruction for binary
syscalls. To bypass binary syscalls, we extract the `__kernel_vsyscall`
symbol from VDSO and replace the `syscall` instruction with a direct
call to the kernel functions.

Note that during the initialization process of glibc, certain syscalls
will be used. However, at this point, the `__kernel_vsyscall` symbol
has not yet been extracted. It's also possible that the kernel could
not provide this interface. In such cases, the `__kernel_vsyscall`
pointer remains at its default value of `NULL`, and glibc will continue
to use binary syscalls.

Since we do not support syscalls with cancellation points (syscall_cp),
we merely map them to regular syscalls.

Signed-off-by: Tianyi Liu <i.pear@outlook.com>
Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Co-authored-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Ioan-Teodor Teugea <ioan_teodor.teugea@stud.acs.upb.ro>
Reviewed-by: Andra Paraschiv <andra@unikraft.org>
Approved-by: Simon Kuenzer <simon@unikraft.io>
GitHub-Closes: #1